+++ examples/storybook/stories/Components/EditorWrapper.js
@@ -158,16 +158,8 @@ class EditorWrapper extends React.Component {
     : plugins;
 
   render() {
-    const {
-      content,
-      palette,
-      config,
-      onChange,
-      isMobile,
-      toolbarSettings,
-      onBlur,
-      onFocus,
-    } = this.props;
+    const { content, palette, onChange, isMobile, toolbarSettings, onBlur, onFocus } = this.props;
+
     return (
       <RicosEditor
         ref={ref => (this.editor = ref)}
@@ -179,12 +171,7 @@ class EditorWrapper extends React.Component {
         toolbarSettings={toolbarSettings}
         onChange={onChange}
       >
-        <RichContentEditor
-          helpers={{ onFilesChange }}
-          config={config}
-          onBlur={onBlur}
-          onFocus={onFocus}
-        />
+        <RichContentEditor helpers={{ onFilesChange }} onBlur={onBlur} onFocus={onFocus} />
       </RicosEditor>
     );
   }
@@ -197,7 +184,6 @@ EditorWrapper.propTypes = {
   isMobile: PropTypes.bool,
   pluginsToDisplay: PropTypes.arrayOf(PropTypes.string),
   toolbarSettings: PropTypes.object,
-  config: PropTypes.object,
   onBlur: PropTypes.func,
   onFocus: PropTypes.func,
 };