+++ packages/ricos-viewer/web/src/RicosViewer.tsx
@@ -5,19 +5,39 @@ import RicosModal from './modals/RicosModal';
 import './styles.css';
 import { RicosViewerProps } from './index';
 
-export class RicosViewer extends Component<RicosViewerProps> {
+interface State {
+  isPreviewExpanded: boolean;
+}
+
+export class RicosViewer extends Component<RicosViewerProps, State> {
+  constructor(props: RicosViewerProps) {
+    super(props);
+    this.state = {
+      isPreviewExpanded: false,
+    };
+  }
+  onPreviewExpand = () => this.setState({ isPreviewExpanded: true });
   render() {
-    const { children, ...props } = this.props;
+    const { children, seoSettings, ...props } = this.props;
+    const { isPreviewExpanded } = this.state;
     const child =
       children && shouldRenderChild('RichContentViewer', children) ? (
         children
       ) : (
         <RichContentViewer />
       );
-
     return (
-      <RicosEngine RicosModal={RicosModal} isViewer key={'viewer'} {...props}>
-        {child}
+      <RicosEngine
+        RicosModal={RicosModal}
+        isPreviewExpanded={isPreviewExpanded}
+        onPreviewExpand={this.onPreviewExpand}
+        isViewer
+        key={'viewer'}
+        {...props}
+      >
+        {React.cloneElement(child, {
+          seoMode: seoSettings,
+        })}
       </RicosEngine>
     );
   }